+++ src/sasctl/core.py
@@ -179,9 +179,9 @@ class Session(requests.Session):
 
     Parameters
     ----------
-    host : str or swat.CAS
+    hostname : str or swat.CAS
         Name of the server to connect to or an established swat.CAS session.
-    user : str, optional
+    username : str, optional
         Username for authentication.  Not required if `host` is a CAS
         connection or if Kerberos is used.  If using Kerberos and an explicit
         username is desired, maybe be a string in 'user@REALM' format.
@@ -203,12 +203,11 @@ class Session(requests.Session):
     Attributes
     ----------
     message_log
-    settings
     filters
 
     """
-    def __init__(self, host,
-                 user=None,
+    def __init__(self, hostname,
+                 username=None,
                  password=None,
                  authinfo=None,
                  protocol=None,
@@ -233,9 +232,14 @@ class Session(requests.Session):
 
         # If certificate path hasn't been specified in either environment
         # variable, replace the default adapter with one that will use the
-        # machine's default SSL settings.
-        if 'REQUESTS_CA_BUNDLE' not in os.environ and verify_ssl:
-            self.mount('https://', SSLContextAdapter())
+        # machine's default SSL _settings.
+        if 'REQUESTS_CA_BUNDLE' not in os.environ:
+            if verify_ssl:
+                self.mount('https://', SSLContextAdapter())
+            else:
+                # Every request will generate an InsecureRequestWarning
+                from urllib3.exceptions import InsecureRequestWarning
+                warnings.simplefilter('default', InsecureRequestWarning)
 
         self.filters = DEFAULT_FILTERS
 
@@ -243,58 +247,62 @@ class Session(requests.Session):
         self._old_session = None
 
         # Reuse an existing CAS connection if possible
-        if swat and isinstance(host, swat.CAS):
-            if isinstance(host._sw_connection,
+        if swat and isinstance(hostname, swat.CAS):
+            if isinstance(hostname._sw_connection,
                           swat.cas.rest.connection.REST_CASConnection):
                 import base64
 
                 # Use the httpAddress action to retieve information about REST endpoints
-                httpAddress = host.get_action('builtins.httpAddress')
+                httpAddress = hostname.get_action('builtins.httpAddress')
                 address = httpAddress()
-
                 domain = address.virtualHost
+                # httpAddress action may return virtualHost = ''
+                # if this happens, try the CAS host
+                if not domain:
+                    domain = hostname._sw_connection._current_hostname
                 protocol = address.protocol
                 port = address.port
 
-                # protocol = host._protocol
-                auth = host._sw_connection._auth.decode('utf-8').replace(
+                # protocol = hostname._protocol
+                auth = hostname._sw_connection._auth.decode('utf-8').replace(
                     'Basic ', '')
-                user, password = base64.b64decode(auth).decode('utf-8').split(
+                username, password = base64.b64decode(auth).decode('utf-8').split(
                     ':')
-                # domain = host._hostname
+                # domain = hostname._hostname
             else:
                 raise ValueError(
                     "A 'swat.CAS' session can only be reused when it's connected via the REST APIs.")
         else:
-            domain = str(host)
+            domain = str(hostname)
 
-        self.settings = {'protocol': protocol or 'https',
+        self._settings = {'protocol': protocol or 'https',
                          'domain': domain,
                          'port': port,
-                         'username': user,
+                         'username': username,
                          'password': password
-                         }
+                          }
 
-        if self.settings['password'] is None:
+        if self._settings['password'] is None:
             # Try to get credentials from .authinfo or .netrc files.
             # If no file path was specified, the default locations will
             # be checked.
             if 'swat' in sys.modules:
-                auth = swat.utils.authinfo.query_authinfo(domain, user=user,
+                auth = swat.utils.authinfo.query_authinfo(domain, user=username,
                                                           path=authinfo)
-                self.settings['username'], self.settings[
-                    'password'] = auth.get('user'), auth.get('password')
+                self._settings['username'], self._settings[
+                    'password'] = auth.get('username'), auth.get('password')
 
             # Not able to load credentials using SWAT.  Try Netrc.
-            # TODO: IF a user was specified, verify that the credentials found
-            #       are for that user.
-            if self.settings['password'] is None:
+            # TODO: IF a username was specified, verify that the credentials found
+            #       are for that username.
+            if self._settings['password'] is None:
                 try:
                     parser = netrc.netrc(authinfo)
                     values = parser.authenticators(domain)
                     if values:
-                        self.settings['username'], _, self.settings[
-                            'password'] = values
+                        self._settings['username'], \
+                        _, \
+                        self._settings['password'] = values
                 except (OSError, IOError):
                     pass  # netrc throws if $HOME is not set
 
@@ -312,8 +320,8 @@ class Session(requests.Session):
         return handler
 
     @property
-    def user(self):
-        return self.settings.get('username')
+    def username(self):
+        return self._settings.get('username')
 
     def send(self, request, **kwargs):
         if self.message_log.isEnabledFor(logging.DEBUG):
@@ -395,11 +403,11 @@ class Session(requests.Session):
             raise RuntimeError("Kerberos package not found.  Run 'pip "
                                "install sasctl[kerberos]' to install.")
 
-        user = self.settings.get('username')
+        user = self._settings.get('username')
         # realm = user.rsplit('@', maxsplit=1)[-1] if '@' in user else None
         client_id = 'sas.tkmtrb'
         flags = kerberos.GSS_C_MUTUAL_FLAG | kerberos.GSS_C_SEQUENCE_FLAG
-        service = 'HTTP@%s' % self.settings['domain']
+        service = 'HTTP@%s' % self._settings['domain']
 
         logger.info('Attempting Kerberos authentication to %s as %s'
                     % (service, user))
@@ -444,7 +452,7 @@ class Session(requests.Session):
 
         # Drop @REALM from username and store
         if username is not None:
-            self.settings['username'] = username.rsplit('@', maxsplit=1)[0]
+            self._settings['username'] = username.rsplit('@', maxsplit=1)[0]
 
 
 
@@ -468,8 +476,8 @@ class Session(requests.Session):
 
     def _get_token_with_password(self):
         """Authenticate with a username and password."""
-        username = self.settings['username']
-        password = self.settings['password']
+        username = self._settings['username']
+        password = self._settings['password']
         url = self._build_url('/SASLogon/oauth/token')
 
         data = 'grant_type=password&username={}&password={}'.format(username,
@@ -481,13 +489,31 @@ class Session(requests.Session):
                                       data=data,
                                       headers=headers,
                                       auth=('sas.ec', ''))
-        r.raise_for_status()
+
+        if r.status_code == 401:
+            raise AuthenticationError("Authentication failed for user '%s'." % username)
+        else:
+            r.raise_for_status()
 
         return r.json().get('access_token')
 
     def get_token(self):
-        username = self.settings['username']
-        password = self.settings['password']
+        """Authenticates with the session host and retrieves an
+        authorization token for use by subsequent requests.
+
+        Returns
+        -------
+        str
+            a bearer token for :class:`HTTPBearerAuth`
+
+        Raises
+        ------
+        AuthenticationError
+            authentication with the host failed
+
+        """
+        username = self._settings['username']
+        password = self._settings['password']
 
         if username is None or password is None:
             return self._get_token_with_kerberos()
@@ -498,14 +524,14 @@ class Session(requests.Session):
         """Build a complete URL from a path by substituting in session parameters."""
         components = urlsplit(url)
 
-        domain = components.netloc or self.settings['domain']
+        domain = components.netloc or self._settings['domain']
 
         # Add port if a non-standard port was specified
-        if self.settings['port'] is not None and ':' not in domain:
-            domain = '{}:{}'.format(domain, self.settings['port'])
+        if self._settings['port'] is not None and ':' not in domain:
+            domain = '{}:{}'.format(domain, self._settings['port'])
 
         return urlunsplit([
-            components.scheme or self.settings['protocol'],
+            components.scheme or self._settings['protocol'],
             domain,
             components.path,
             components.query,
@@ -520,21 +546,12 @@ class Session(requests.Session):
         self._old_session = _session
         _session = self
 
-        # Reduce verbosity of warnings.  By default, HTTP warnings on thrown on every request
-        self._filter_context = warnings.catch_warnings()
-        self._filter_context.__enter__()
-        from urllib3.exceptions import HTTPWarning
-        warnings.simplefilter('once', HTTPWarning)
-
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         # Restore previous current session
         global _session
         _session = self._old_session
-
-        # Restore warning levels
-        self._filter_context.__exit__()
 
         super(Session, self).__exit__()
 
@@ -639,6 +656,10 @@ def request_link(obj, rel, **kwargs):
 
     """
     link = get_link(obj, rel)
+
+    if link is None:
+        raise ValueError("Link '%s' not found in object %s." % (rel, obj))
+
 
     return request(link['method'], link['href'], **kwargs)
 
@@ -874,3 +895,13 @@ def _build_is_available_func(service_root):
         response = current_session().head(service_root + '/')
         return response.status_code == 200
     return is_available
+
+
+class SasctlError(Exception):
+    pass
+
+class TimeoutError(SasctlError):
+    pass
+
+class AuthenticationError(SasctlError):
+    pass