+++ tests/integration/test_files.py
@@ -5,9 +5,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import pytest
+from six.moves import mock
 
 from sasctl import RestObj
-
+from sasctl.services import files
 
 pytestmark = pytest.mark.usefixtures('session')
 
@@ -21,10 +22,9 @@ def dummy_file(tmpdir):
 
 
 def test_list_files():
-    from sasctl.services.files import list_files
-    files = list_files()
+    all_files = files.list_files()
 
-    assert all(isinstance(f, RestObj) for f in files)
+    assert all(isinstance(f, RestObj) for f in all_files)
 
 
 @pytest.mark.incremental
@@ -32,97 +32,95 @@ class TestFile:
     filename = 'sasctl_test_file'
 
     def test_create_file_with_name(self, dummy_file):
-        """Create a file with an explicitly set filename."""
-        from sasctl.services.files import create_file
+        # Requests uses os.urandom(16) to generate boundaries for multi-part
+        # form uploads.  Mock the output to ensure a consistent value so
+        # that body request/responses always match.
+        with mock.patch('os.urandom',
+                        return_value='abcdefghijklmnop'.encode('utf-8')):
 
-        file = create_file(dummy_file, filename=self.filename)
+            """Create a file with an explicitly set filename."""
+            file = files.create_file(dummy_file, filename=self.filename)
 
         assert isinstance(file, RestObj)
 
     def test_get_file_with_name(self):
         """Ensure previously created file can be retrieved."""
-        from sasctl.services.files import get_file
-
-        file = get_file(self.filename)
+        file = files.get_file(self.filename)
 
         assert isinstance(file, RestObj)
         assert self.filename == file.name
 
     def test_get_file_content(self, dummy_file):
-        from sasctl.services.files import get_file_content
 
         with open(dummy_file, 'r') as f:
             target = f.read()
 
-        content = get_file_content(self.filename)
+        content = files.get_file_content(self.filename)
 
         assert target == content
 
     def test_delete_file_with_name(self):
         """Delete previously created file."""
-        from sasctl.services.files import get_file, delete_file
 
-        delete_file(self.filename)
-        file = get_file(self.filename)
+        files.delete_file(self.filename)
+        file = files.get_file(self.filename)
 
         assert file is None
 
     def test_create_file_without_name(self, dummy_file):
         """Create a file from just a path."""
-        from sasctl.services.files import create_file
-
-        file = create_file(dummy_file)
+        # Requests uses os.urandom(16) to generate boundaries for multi-part
+        # form uploads.  Mock the output to ensure a consistent value so
+        # that body request/responses always match.
+        with mock.patch('os.urandom',
+                        return_value='abcdefghijklmnop'.encode('utf-8')):
+            file = files.create_file(dummy_file)
 
         assert isinstance(file, RestObj)
         assert FILENAME == file.name
 
     def test_get_file_without_name(self):
         """Ensure previously created file can be retrieved."""
-        from sasctl.services.files import get_file
-
-        file = get_file(FILENAME)
+        file = files.get_file(FILENAME)
 
         assert isinstance(file, RestObj)
         assert FILENAME == file.name
 
     def test_delete_file_without_name(self):
         """Delete previously created file."""
-        from sasctl.services.files import get_file, delete_file
-
-        delete_file(FILENAME)
-        file = get_file(FILENAME)
+        files.delete_file(FILENAME)
+        file = files.get_file(FILENAME)
 
         assert file is None
 
     def test_create_file_from_file_object(self, dummy_file):
         """Create a file from just a path."""
-        from sasctl.services.files import create_file
-
 
         with open(dummy_file) as f:
             # Filename must be provided
             with pytest.raises(ValueError):
-                file = create_file(f)
+                file = files.create_file(f)
 
-            file = create_file(f, filename=self.filename)
+            # Requests uses os.urandom(16) to generate boundaries for multi-part
+            # form uploads.  Mock the output to ensure a consistent value so
+            # that body request/responses always match.
+            with mock.patch('os.urandom',
+                            return_value='abcdefghijklmnop'.encode('utf-8')):
+                file = files.create_file(f, filename=self.filename)
 
         assert isinstance(file, RestObj)
         assert self.filename == file.name
 
     def test_get_file_from_file_object(self):
         """Ensure previously created file can be retrieved."""
-        from sasctl.services.files import get_file
-
-        file = get_file(self.filename)
+        file = files.get_file(self.filename)
 
         assert isinstance(file, RestObj)
         assert self.filename == file.name
 
     def test_delete_file_without_name(self):
         """Delete previously created file."""
-        from sasctl.services.files import get_file, delete_file
-
-        delete_file(self.filename)
-        file = get_file(self.filename)
+        files.delete_file(self.filename)
+        file = files.get_file(self.filename)
 
         assert file is None