+++ tests/unit/test_microanalytic_score.py
@@ -7,9 +7,10 @@
 import pytest
 from six.moves import mock
 
+from sasctl.services import microanalytic_score as mas
 
 def test_create_python_module():
-    from sasctl.services.microanalytic_score import create_module
+
     from sasctl import current_session
 
     with mock.patch('sasctl.core.requests.Session.request'):
@@ -17,7 +18,7 @@ def test_create_python_module():
 
     with mock.patch('sasctl.services.microanalytic_score.post') as post:
         with pytest.raises(ValueError):
-            create_module()     # Source code is required
+            mas.create_module()     # Source code is required
 
     with mock.patch('sasctl.services.microanalytic_score.post') as post:
         source = '\n'.join(("def testMethod(var1, var2):",
@@ -25,7 +26,7 @@ def test_create_python_module():
                             "    out1 = var1 + 5",
                             "    out2 = var2.upper()",
                             "    return out1, out2"))
-        create_module(source=source)
+        mas.create_module(source=source)
 
         assert post.call_count == 1
         json = post.call_args[1].get('json', {})